Rewriting lab (#6)
* fixed logic to make more independence

* fixed problem with harmony calculation

* decided to create a hackathon factory + moved calc avg harmony to hackathon

* added tests to hrmanager & director

* renamed value to understandable name + added tests to director

* improved calc harmony mean method

* added test for hackathon, minifixes to wishlist generator

* moved strategy (added interface) to write correct test

* decomposed file loading + fixed tests with strategy

* fixed remarks + added data loader tests

* added validation of options

* added more exceptions and tests

* added case with duplicate id

+++ Nsu.Sharps.Hackathon.NetGenericHost/Nsu.Sharps.Hackathon.NetGenericHost/Options/CalculationDataOptions.cs
@@ -2,6 +2,6 @@ namespace Nsu.Sharps.Hackathon.NetGenericHost.Options;
 
 public class CalculationDataOptions
 {
-    public double NumeratorInCountingHarmony { get; set; }
-    public double AmountOfParticipantsInNumerator { get; set; }
+    public int NumeratorInCountingHarmony { get; set; }
+    public int AmountOfParticipantsInNumerator { get; set; }
 }
